Refactor photo component
