+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { default as MenuButton } from './Buttons';
+export { default as Modal } from './Modal';
+export { default as Navbar } from './Navbar';
