+++ src/index.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Container } from 'styles';
 import { GlobalStyles } from 'styles/GlobalStyles';
+import 'react-loader-spinner/dist/loader/css/react-spinner-loader.css';
 import App from './App';
 
 ReactDOM.render(
